+++ lib/strong_migrations/error_messages.rb
@@ -166,8 +166,8 @@ Instead, start an index with columns that narrow down the results the most.",
 Upgrade Postgres before adding new indexes, or wrap this step in a safety_assured { ... } block
 to accept the risk.",
 
-    add_index_copy:
-"Adding an index with copying blocks writes. Instead, use:
+    copy_algorithm:
+"Using the COPY algorithm blocks writes. Instead, use:
 
 class %{migration_name} < ActiveRecord::Migration%{migration_suffix}
   def change